+++ changelog.md
@@ -1,3 +1,4 @@
 - Added Simplified Chinese translation
 - Added show empty containers config option
 - Added Shulker Box item hint
+- Fixed honey level not displayed in all cases